requirements.txt
streamlit~=1.44.1
boto3~=1.37.28
pandas~=2.2.3
plotly~=6.0.1
mutagen~=1.47.0
python-dotenv~=1.1.0
sqlalchemy~=2.0.40
pymysql~=1.1.1